backend/requirements.txt
fastapi
uvicorn
google-generativeai
python-owasp-zap-v2.4
requests
pydantic
httpx
multiprocess
